+++ core/src/db/meta/condition/MetaCombination.h
@@ -1,0 +1,49 @@
+// Copyright (C) 2019-2020 Zilliz. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance
+// with the License. You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software distributed under the License
+// is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+// or implied. See the License for the specific language governing permissions and limitations under the License.
+
+#pragma once
+
+#include <string>
+
+#include "db/meta/condition/MetaBaseCombination.h"
+#include "db/meta/condition/MetaBaseFilter.h"
+
+namespace milvus::engine::meta {
+
+class MetaFilterCombination : public MetaBaseCombination {
+ public:
+    explicit MetaFilterCombination(MetaFilterPtr filter) : MetaBaseCombination(one_), filter_(filter) {
+    }
+
+    ~MetaFilterCombination() override = default;
+
+    std::string
+    Dump() const override;
+
+ private:
+    MetaFilterPtr filter_;
+};
+
+class MetaRelationCombination : public MetaBaseCombination {
+ public:
+    MetaRelationCombination(Comb comb, MetaConditionPtr lcond, MetaConditionPtr rcond);
+
+    ~MetaRelationCombination() override = default;
+
+    std::string
+    Dump() const override;
+
+ private:
+    MetaConditionPtr lcond_;
+    MetaConditionPtr rcond_;
+};
+
+}  // namespace milvus::engine::meta
