+++ core/src/db/meta/condition/MetaRelation.cpp
@@ -1,3 +1,4 @@
+
 // Copyright (C) 2019-2020 Zilliz. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance
@@ -9,7 +10,11 @@
 // is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 // or implied. See the License for the specific language governing permissions and limitations under the License.
 
-#include "db/meta/MetaFilter.h"
+#include "db/meta/condition/MetaRelation.h"
+
+#include <memory>
+
+#include "db/meta/condition/MetaCombination.h"
 
 namespace milvus::engine::meta {
